drawTileGrid function now properly draws an isometric tilegrid with (i think) correct 2D plane -> 3D coordinates. Batching now works with textures for drawing tile textures.

+++ IsoEngineEditor/src/IsoEditorLayer.h
@@ -23,7 +23,7 @@ namespace IE
 		void OnUpdate(Timestep timestep) override;
 		void OnEvent(Event& ev) override;
 
-		void CreateTileGrid(const glm::uvec2& gridSize, const glm::uvec2& tileSize);
+		void CreateTileGrid(const glm::uvec2& gridSize, const glm::vec2& tileSize);
 
 	private:
 		void NewScene();
@@ -53,7 +53,7 @@ namespace IE
 
 		/* Tile Grid */
 		glm::uvec2 m_TileGridSize;
-		glm::uvec2 m_TileSize;
+		glm::vec2 m_TileSize;
 		glm::vec2 m_Origin;
 	};
